requirements.txt
flask
tensorflow
transformers
torch
pytesseract
pillow
